+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## **v1.1.0** - 2025-04-08
+### ➕ Added
+* [web-audio] `autoplaySounds` options
+* [web-audio] `ready` event
+
+## **v1.0.3** - 2025-01-07
+### ➕ Added
+* manual spot to web example
+
+### 🪛 Changed
+* improved documentation for manual mode
+
+### 🩹 Fixed
+* make spot position optional to enable manual spots configuration
+
 ## **v1.0.2** - 2024-12-23
 ### 🪛 Changed
 * refine README